refactor: :recycle: DateWrapper 에 있던 partitionByMonth 함수를 cursus user service 안에서만 사용할 수 있도록 변경
해당 함수가 범용성도 없고, 오류를 발생시키기 너무 쉽다고 판단하여,
지금 필요로 하는 유일한 함수인 aliveUserCountRecords 에서만 사용하도록 cursus user service 로 옮겼습니다.
